Update README.md
In Pokémon Detective Pikachu, a detective named Harry Goodman goes missing and his son Tim is on the case to figure out what happened. Filled with eye-popping visuals and Ryan Reynolds’ signature humor, the film is a rollicking ride of action, humor, and heart, pitting a wisecracking, mystery-solving detective against a suspiciously sinister plot in the underground city of Ryme City.